12.3 生成车次时，同步生成余票信息

+++ business/src/main/java/com/miker/train/business/domain/DailyTrainTicket.java
@@ -16,7 +16,7 @@ public class DailyTrainTicket {
 
     private Date startTime;
 
-    private Byte startIndex;
+    private Integer startIndex;
 
     private String end;
 
@@ -24,7 +24,7 @@ public class DailyTrainTicket {
 
     private Date endTime;
 
-    private Byte endIndex;
+    private Integer endIndex;
 
     private Integer ydz;
 
@@ -94,11 +94,11 @@ public class DailyTrainTicket {
         this.startTime = startTime;
     }
 
-    public Byte getStartIndex() {
+    public Integer getStartIndex() {
         return startIndex;
     }
 
-    public void setStartIndex(Byte startIndex) {
+    public void setStartIndex(Integer startIndex) {
         this.startIndex = startIndex;
     }
 
@@ -126,11 +126,11 @@ public class DailyTrainTicket {
         this.endTime = endTime;
     }
 
-    public Byte getEndIndex() {
+    public Integer getEndIndex() {
         return endIndex;
     }
 
-    public void setEndIndex(Byte endIndex) {
+    public void setEndIndex(Integer endIndex) {
         this.endIndex = endIndex;
     }
