+++ business/src/main/java/com/miker/train/business/domain/DailyTrainTicketExample.java
@@ -559,52 +559,52 @@ public class DailyTrainTicketExample {
             return (Criteria) this;
         }
 
-        public Criteria andStartIndexEqualTo(Byte value) {
+        public Criteria andStartIndexEqualTo(Integer value) {
             addCriterion("start_index =", value, "startIndex");
             return (Criteria) this;
         }
 
-        public Criteria andStartIndexNotEqualTo(Byte value) {
+        public Criteria andStartIndexNotEqualTo(Integer value) {
             addCriterion("start_index <>", value, "startIndex");
             return (Criteria) this;
         }
 
-        public Criteria andStartIndexGreaterThan(Byte value) {
+        public Criteria andStartIndexGreaterThan(Integer value) {
             addCriterion("start_index >", value, "startIndex");
             return (Criteria) this;
         }
 
-        public Criteria andStartIndexGreaterThanOrEqualTo(Byte value) {
+        public Criteria andStartIndexGreaterThanOrEqualTo(Integer value) {
             addCriterion("start_index >=", value, "startIndex");
             return (Criteria) this;
         }
 
-        public Criteria andStartIndexLessThan(Byte value) {
+        public Criteria andStartIndexLessThan(Integer value) {
             addCriterion("start_index <", value, "startIndex");
             return (Criteria) this;
         }
 
-        public Criteria andStartIndexLessThanOrEqualTo(Byte value) {
+        public Criteria andStartIndexLessThanOrEqualTo(Integer value) {
             addCriterion("start_index <=", value, "startIndex");
             return (Criteria) this;
         }
 
-        public Criteria andStartIndexIn(List<Byte> values) {
+        public Criteria andStartIndexIn(List<Integer> values) {
             addCriterion("start_index in", values, "startIndex");
             return (Criteria) this;
         }
 
-        public Criteria andStartIndexNotIn(List<Byte> values) {
+        public Criteria andStartIndexNotIn(List<Integer> values) {
             addCriterion("start_index not in", values, "startIndex");
             return (Criteria) this;
         }
 
-        public Criteria andStartIndexBetween(Byte value1, Byte value2) {
+        public Criteria andStartIndexBetween(Integer value1, Integer value2) {
             addCriterion("start_index between", value1, value2, "startIndex");
             return (Criteria) this;
         }
 
-        public Criteria andStartIndexNotBetween(Byte value1, Byte value2) {
+        public Criteria andStartIndexNotBetween(Integer value1, Integer value2) {
             addCriterion("start_index not between", value1, value2, "startIndex");
             return (Criteria) this;
         }
@@ -819,52 +819,52 @@ public class DailyTrainTicketExample {
             return (Criteria) this;
         }
 
-        public Criteria andEndIndexEqualTo(Byte value) {
+        public Criteria andEndIndexEqualTo(Integer value) {
             addCriterion("end_index =", value, "endIndex");
             return (Criteria) this;
         }
 
-        public Criteria andEndIndexNotEqualTo(Byte value) {
+        public Criteria andEndIndexNotEqualTo(Integer value) {
             addCriterion("end_index <>", value, "endIndex");
             return (Criteria) this;
         }
 
-        public Criteria andEndIndexGreaterThan(Byte value) {
+        public Criteria andEndIndexGreaterThan(Integer value) {
             addCriterion("end_index >", value, "endIndex");
             return (Criteria) this;
         }
 
-        public Criteria andEndIndexGreaterThanOrEqualTo(Byte value) {
+        public Criteria andEndIndexGreaterThanOrEqualTo(Integer value) {
             addCriterion("end_index >=", value, "endIndex");
             return (Criteria) this;
         }
 
-        public Criteria andEndIndexLessThan(Byte value) {
+        public Criteria andEndIndexLessThan(Integer value) {
             addCriterion("end_index <", value, "endIndex");
             return (Criteria) this;
         }
 
-        public Criteria andEndIndexLessThanOrEqualTo(Byte value) {
+        public Criteria andEndIndexLessThanOrEqualTo(Integer value) {
             addCriterion("end_index <=", value, "endIndex");
             return (Criteria) this;
         }
 
-        public Criteria andEndIndexIn(List<Byte> values) {
+        public Criteria andEndIndexIn(List<Integer> values) {
             addCriterion("end_index in", values, "endIndex");
             return (Criteria) this;
         }
 
-        public Criteria andEndIndexNotIn(List<Byte> values) {
+        public Criteria andEndIndexNotIn(List<Integer> values) {
             addCriterion("end_index not in", values, "endIndex");
             return (Criteria) this;
         }
 
-        public Criteria andEndIndexBetween(Byte value1, Byte value2) {
+        public Criteria andEndIndexBetween(Integer value1, Integer value2) {
             addCriterion("end_index between", value1, value2, "endIndex");
             return (Criteria) this;
         }
 
-        public Criteria andEndIndexNotBetween(Byte value1, Byte value2) {
+        public Criteria andEndIndexNotBetween(Integer value1, Integer value2) {
             addCriterion("end_index not between", value1, value2, "endIndex");
             return (Criteria) this;
         }
